CORE-9839 move signing out of uniqueness client service
