+++ gradle.properties
@@ -40,7 +40,7 @@ bouncycastleVersion=1.72
 # Corda API libs revision (change in 4th digit indicates a breaking change)
 # Change to 5.0.0.xx-SNAPSHOT to pick up maven local published copy
 #cordaApiVersion=5.0.0.xxx-SNAPSHOT
-cordaApiVersion=5.0.0.629-beta+
+cordaApiVersion=5.0.0.631-alpha-1675339104618
 
 disruptorVersion=3.4.2
 felixConfigAdminVersion=1.9.26
