--- libs/uniqueness/common/src/main/kotlin/net/corda/uniqueness/datamodel/impl/UniquenessCheckResponseImpl.kt
@@ -1,10 +1,0 @@
-package net.corda.uniqueness.datamodel.impl
-
-import net.corda.v5.application.crypto.DigitalSignatureAndMetadata
-import net.corda.v5.application.uniqueness.model.UniquenessCheckResult
-import net.corda.v5.ledger.utxo.uniqueness.client.LedgerUniquenessCheckResponse
-
-class UniquenessCheckResponseImpl(
-    override val result: UniquenessCheckResult,
-    override val signature: DigitalSignatureAndMetadata?
-) : LedgerUniquenessCheckResponse
