+++ notary-plugins/notary-plugin-non-validating/notary-plugin-non-validating-server/src/main/kotlin/com/r3/corda/notary/plugin/nonvalidating/server/NonValidatingNotaryServerFlowImpl.kt
@@ -6,17 +6,29 @@ import com.r3.corda.notary.plugin.common.NotaryErrorGeneralImpl
 import com.r3.corda.notary.plugin.common.toNotarisationResponse
 import com.r3.corda.notary.plugin.common.validateRequestSignature
 import com.r3.corda.notary.plugin.nonvalidating.api.NonValidatingNotarisationPayload
+import net.corda.v5.application.crypto.DigestService
+import net.corda.v5.application.crypto.DigitalSignatureAndMetadata
+import net.corda.v5.application.crypto.DigitalSignatureMetadata
 import net.corda.v5.application.crypto.DigitalSignatureVerificationService
+import net.corda.v5.application.crypto.MerkleTreeFactory
+import net.corda.v5.application.crypto.SigningService
 import net.corda.v5.application.flows.CordaInject
 import net.corda.v5.application.flows.InitiatedBy
 import net.corda.v5.application.flows.ResponderFlow
 import net.corda.v5.application.membership.MemberLookup
 import net.corda.v5.application.messaging.FlowSession
 import net.corda.v5.application.serialization.SerializationService
+import net.corda.v5.application.uniqueness.model.UniquenessCheckResultSuccess
 import net.corda.v5.base.annotations.Suspendable
 import net.corda.v5.base.annotations.VisibleForTesting
 import net.corda.v5.base.util.debug
 import net.corda.v5.base.util.trace
+import net.corda.v5.crypto.DigestAlgorithmName
+import net.corda.v5.crypto.SecureHash
+import net.corda.v5.crypto.SignatureSpec
+import net.corda.v5.crypto.containsAny
+import net.corda.v5.crypto.merkle.HASH_DIGEST_PROVIDER_DEFAULT_NAME
+import net.corda.v5.crypto.merkle.MerkleTree
 import net.corda.v5.ledger.common.Party
 import net.corda.v5.ledger.utxo.StateAndRef
 import net.corda.v5.ledger.utxo.StateRef
@@ -25,6 +37,8 @@ import net.corda.v5.ledger.utxo.transaction.filtered.UtxoFilteredTransaction
 import net.corda.v5.ledger.utxo.uniqueness.client.LedgerUniquenessCheckerClientService
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+import java.security.PublicKey
+import java.time.Instant
 
 /**
  * The server-side implementation of the non-validating notary logic.
@@ -49,6 +63,15 @@ class NonValidatingNotaryServerFlowImpl() : ResponderFlow {
 
     @CordaInject
     private lateinit var memberLookup: MemberLookup
+
+    @CordaInject
+    private lateinit var digestService: DigestService
+
+    @CordaInject
+    private lateinit var signingService: SigningService
+
+    @CordaInject
+    private lateinit var merkleTreeFactory: MerkleTreeFactory
 
     /**
      * Constructor used for testing to initialize the necessary services
@@ -106,28 +129,35 @@ class NonValidatingNotaryServerFlowImpl() : ResponderFlow {
 
             logger.trace { "Requesting uniqueness check for transaction ${txDetails.id}" }
 
-            val uniquenessResponse = clientService.requestUniquenessCheck(
+            val uniquenessResult = clientService.requestUniquenessCheck(
                 txDetails.id.toString(),
                 txDetails.inputs.map { it.toString() },
                 txDetails.references.map { it.toString() },
                 txDetails.numOutputs,
                 txDetails.timeWindow.from,
-                txDetails.timeWindow.until,
-                requestPayload.notaryKey
+                txDetails.timeWindow.until
             )
 
             logger.debug {
-                "Uniqueness check completed for transaction ${txDetails.id}, result is: ${uniquenessResponse.result}. Sending response " +
+                "Uniqueness check completed for transaction ${txDetails.id}, result is: ${uniquenessResult}. Sending response " +
                         "to ${session.counterparty}"
             }
 
-            session.send(uniquenessResponse.toNotarisationResponse())
+            val signature = if (uniquenessResult is UniquenessCheckResultSuccess) {
+                signBatch(listOf(txDetails.id), requestPayload.notaryKey).rootSignature
+            } else null
+
+            val notarisationResponse = uniquenessResult.toNotarisationResponse(signature)
+
+            session.send(notarisationResponse)
         } catch (e: Exception) {
             logger.warn("Error while processing request from client. Cause: $e")
-            session.send(NotarisationResponse(
-                emptyList(),
-                NotaryErrorGeneralImpl("Error while processing request from client.", e)
-            ))
+            session.send(
+                NotarisationResponse(
+                    emptyList(),
+                    NotaryErrorGeneralImpl("Error while processing request from client.", e)
+                )
+            )
         }
     }
 
@@ -211,4 +241,87 @@ class NonValidatingNotaryServerFlowImpl() : ResponderFlow {
 
     private inline fun <reified T> Any.castOrThrow(error: () -> String) = this as? T
         ?: throw java.lang.IllegalStateException(error())
+
+    @Suspendable
+    private fun signBatch(txIds: List<SecureHash>, notaryServiceKey: PublicKey): BatchSignature {
+        // TODO CORE-6615 This validation mechanism needs to be
+        //  reconsidered in the future
+        val algorithms = txIds.mapTo(HashSet(), SecureHash::algorithm)
+        require(algorithms.size > 0) {
+            "Cannot sign an empty batch"
+        }
+        require(algorithms.size == 1) {
+            "Cannot sign a batch with multiple hash algorithms: $algorithms"
+        }
+
+        val algorithm = algorithms.first()
+
+        val allLeaves = txIds.map {
+            // we don't have a reHash function anymore
+            digestService.hash(it.bytes, DigestAlgorithmName(algorithm))
+        }
+
+        val hashDigest = merkleTreeFactory.createHashDigest(
+            HASH_DIGEST_PROVIDER_DEFAULT_NAME,
+            DigestAlgorithmName(algorithm)
+        )
+
+        val merkleTree = merkleTreeFactory.createTree(
+            allLeaves.map { it.bytes },
+            hashDigest
+        )
+
+        val sig = signingService.sign(
+            merkleTree.root.bytes,
+            selectNotaryVNodeSigningKey(notaryServiceKey),
+            SignatureSpec.ECDSA_SHA256
+        )
+
+        return BatchSignature(
+            DigitalSignatureAndMetadata(
+                sig,
+                DigitalSignatureMetadata(
+                    Instant.now(),
+                    SignatureSpec.ECDSA_SHA256,
+                    mapOf(
+                        "platformVersion" to memberLookup.myInfo().platformVersion.toString()
+                    )
+                )
+            ),
+            merkleTree
+        )
+    }
+
+    // TODO CORE-9469 The key selection here will be replaced with the Crypto API once it is finished.
+    @Suspendable
+    private fun selectNotaryVNodeSigningKey(notaryServiceKey: PublicKey): PublicKey {
+
+        // We select the first key that is not null
+        val selectedSigningKey = signingService
+            .findMySigningKeys(setOf(notaryServiceKey)).values
+            .filterNotNull()
+            .firstOrNull()
+
+        // We need to make sure there's at least one key we can sign with
+        require(selectedSigningKey != null) {
+            "Could not find any keys associated with the notary service's public key."
+        }
+
+        // We double check that the selected key is actually part of the notary service key
+        require(notaryServiceKey.containsAny(setOf(selectedSigningKey))) {
+            "The notary key selected for signing is not associated with the notary service key."
+        }
+
+        return selectedSigningKey
+    }
+}
+
+data class BatchSignature(
+    val rootSignature: DigitalSignatureAndMetadata,
+    val fullMerkleTree: MerkleTree
+) {
+
+    /** Extracts a signature with a partial Merkle tree for the specified leaf in the batch signature. */
+    // TODO CORE-6243 how to do this without partial merkle trees?
+    /*fun forParticipant(txId: SecureHash, hashingService: HashingService) { }*/
 }
